tests now working on mingw
